address review feedback

+++ src/Libraries/Microsoft.Extensions.DataIngestion/Writers/VectorStoreWriter.cs
@@ -62,7 +62,7 @@ public sealed class VectorStoreWriter<T> : IngestionChunkWriter<T>
 
         IReadOnlyList<object>? preExistingKeys = null;
         List<Dictionary<string, object?>>? batch = null;
-        int currentBatchTokenCount = 0;
+        long currentBatchTokenCount = 0;
 
         await foreach (IngestionChunk<T> chunk in chunks.WithCancellation(cancellationToken))
         {
@@ -96,27 +96,23 @@ public sealed class VectorStoreWriter<T> : IngestionChunkWriter<T>
                 }
             }
 
-            // Lazy allocate the batch list
             batch ??= [];
 
             // Check if adding this chunk would exceed the batch token limit
-            // If the batch is empty or the chunk alone exceeds the limit, add it anyway (requirement 8)
+            // If the batch is empty or the chunk alone exceeds the limit, add it anyway.
             if (batch.Count > 0 && currentBatchTokenCount + chunk.TokenCount > _options.BatchTokenCount)
             {
-                // Upsert the current batch
                 await _vectorStoreCollection.UpsertAsync(batch, cancellationToken).ConfigureAwait(false);
 
-                // Clear the batch for reuse
                 batch.Clear();
                 currentBatchTokenCount = 0;
             }
 
-            // Add the current chunk to the batch
             batch.Add(record);
             currentBatchTokenCount += chunk.TokenCount;
         }
 
-        // Upsert any remaining chunks in the batch (requirement 7)
+        // Upsert any remaining chunks in the batch
         if (batch?.Count > 0)
         {
             await _vectorStoreCollection!.UpsertAsync(batch, cancellationToken).ConfigureAwait(false);
